finish improve security when delete and fix miss spelling

+++ trip/views.py
@@ -1,3 +1,4 @@
+from django.contrib import auth
 from django.contrib.auth.models import User
 from django.http import HttpResponseNotFound, HttpResponseRedirect
 from django.urls import reverse, reverse_lazy
@@ -214,12 +215,13 @@ def delete_post(request, pk):
     """
     post = get_object_or_404(TripPlan, id=pk)
     if request.method == "POST":
-        image_path = os.path.join(MEDIA_ROOT, 'pic', str(pk))
-        if os.path.exists(image_path):
-            shutil.rmtree(image_path)
-        post.delete()
-        success_url = reverse_lazy('trip:tripplan')
-        return redirect(success_url)
+        if request.user.id == post.author.id:
+            image_path = os.path.join(MEDIA_ROOT, 'pic', str(pk))
+            if os.path.exists(image_path):
+                shutil.rmtree(image_path)
+            post.delete()
+            success_url = reverse_lazy('trip:tripplan')
+            return redirect(success_url)
     context = {
         "post": post
     }
